alpha: get compilers/zc to build, matching vc's own type-signature fixes
Same class of fix already applied to vc when it was brought up:
- swcmp's gc.h prototype needs the const-qualified qsort-comparator
  signature the shared compilers/cck/pswt.c actually defines
  (int(const void*, const void*)), not zc's own unqualified one.
- align()/maxround() (swt.c) need int32, not long, to match the
  int32-typed prototypes in the shared compilers/cck/cc.h.
- hintabsize needs a plain `extern` declaration in gc.h (matching
  vc/gc.h's own "defined in mul.c" comment), not the EXTERN macro --
  it's a real, initialized definition living solely in mul.c, and the
  EXTERN macro would make list.c (the file that #define EXTERN's away
  the macro to own every other global) provide a second, conflicting
  tentative definition of the same symbol.

Also fixes the pre-rename '7' -> 'z' thechar leftover in txt.c's
ginit() (same leftover class as zl's own, already fixed), which was
otherwise harmless to compile but printed "usage: 7c" instead of "zc".

Verified: zc builds, links, and compiles a trivial C file without
crashing.

+++ compilers/zc/swt.c
@@ -500,10 +500,10 @@ zaddr(char *bp, Adr *a, int s)
 	return bp;
 }
 
-long
-align(long i, Type *t, int op)
+int32
+align(int32 i, Type *t, int op)
 {
-	long o;
+	int32 o;
 	Type *v;
 	int w;
 
@@ -566,8 +566,8 @@ align(long i, Type *t, int op)
 	return o;
 }
 
-long
-maxround(long max, long v)
+int32
+maxround(int32 max, int32 v)
 {
 	v += SZ_VLONG-1;
 	if(v > max)
